Bug 32125: Fix circuit display for bridge without a fingerprint
Torbutton expects all bridges to have a fingerprint when creating
the browser's circuit display. This patch works around the case
when the user provides a bridge without a fingerprint by assuming
it is a bridge, but we cannot determine the other displayed info:
type (obfs4, meek, etc) nor the ip.

In this scenario, the entry node in the circuit display will simply
say "Bridge".